+++ docs/guia/campos_utilizados.md
@@ -1,0 +1,78 @@
+# Guia de Uso
+
+## Campos utilizados para o cálculo
+### Custo do Produto (R$)
+!!! info "O que é?"
+    Representa o custo unitário do produto que você deseja vender. É o valor que você paga para adquirir ou produzir o item.
+
+!!! tip "Dica"
+    - Inclua apenas o custo direto do produto.
+    - Não inclua custos de embalagem ou outros gastos.
+
+---
+
+### Custo da Embalagem (R$)
+!!! info "O que é?"
+    Refere-se ao custo da embalagem necessária para enviar o produto ao cliente.
+
+!!! tip "Dica"
+    - Considere o custo da caixa, plástico bolha, etiquetas, etc.
+    - Se a embalagem for reutilizável, divida o custo pelo número de usos.
+
+---
+
+### Taxa de Imposto (%)
+!!! info "O que é?"
+    É a porcentagem de imposto que incide sobre o preço de venda do produto.
+
+!!! tip "Dica"
+    - Verifique a alíquota de impostos aplicável ao seu produto.
+    - Considere impostos federais, estaduais e municipais.
+
+---
+
+### Taxa da Shopee (%)
+!!! info "O que é?"
+    Representa a comissão que a Shopee cobra sobre cada venda.
+
+!!! tip "Dica"
+    - Consulte a tabela de comissões da Shopee para o seu segmento.
+    - Inclua taxas adicionais, como frete subsidiado, se aplicável.
+    - Para vendedores que possuem programa de frete grátis, o valor é de 20%. Para aqueles que não possuem, o valor é de 14%.
+
+---
+
+### Valor Fixo Shopee (R$)
+!!! info "O que é?"
+    É um valor fixo que a Shopee cobra por transação, independentemente do preço do produto.
+
+!!! tip "Dica"
+    - Verifique na política da Shopee se há cobranças fixas.
+    - Inclua taxas de pagamento, se houver.
+    - Por padrão a shopee cobra R$4,00 para vendedores CNPJ e R$5,00 para CPF. 
+    - Para produtos com valor de venda abaixo de R$8,00 a taxa de acrescimo é de 50% do valor do produto, ou seja, se o produto custa R$7,00, a taxa será de R$3,50
+
+---
+
+### Outros Custos (R$)
+!!! info "O que é?"
+    São custos adicionais que você tem com o produto, como logística, armazenagem ou marketing.
+
+!!! tip "Dica"
+    - Inclua todos os custos indiretos que impactam o preço final.
+    - Se os custos forem compartilhados entre vários produtos, divida o valor proporcionalmente.
+
+---
+
+### Exemplo de Preenchimento
+
+Aqui está um exemplo de como preencher os campos para um produto fictício:
+
+| Campo                  | Valor  |
+|------------------------|--------|
+| Custo do Produto (R$)   | 50.00  |
+| Custo da Embalagem (R$) | 5.00   |
+| Taxa de Imposto (%)     | 12.00  |
+| Taxa da Shopee (%)      | 15.00  |
+| Valor Fixo Shopee (R$)  | 2.00   |
+| Outros Custos (R$)      | 3.00   |
