+++ docs/index.md
@@ -1,0 +1,16 @@
+# Documentação: Calculadora de Preços E-Commerce
+
+Este é um aplicativo para cálculo de preço de venda ideal para anúncios de E-commerce, todo o cálculo é feito com base nos custos do produto, custos operacionais, impostos e taxas de E-commerce.
+
+## Funcionalidades
+
+### **Cálculos para Shopee**:
+
+- **Calculo de produto individual**: Calcule o preço de venda de um único produto.
+
+- **Calculo em Massa**: Calcule o preço de venda para vários produtos de uma só vez.
+
+### **Cálculos para Mercado Livre**:
+- Em desenvolvimento
+
+[Guia de uso](guia)
