+++ src/processar_planilha.py
@@ -1,45 +1,48 @@
+import sys
+import os
 import pandas as pd
-from io import BytesIO
-import streamlit as st
 
-from src import utils
+sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+from classes.CalculoPrecoShopee import CalculoPrecoShopee
 
-# Título do app
-st.title("Calculadora de Preços para Múltiplos Produtos")
-
-# Upload da planilha
-uploaded_file = st.file_uploader("Faça o upload da planilha (Excel ou CSV)", type=["xlsx", "csv"])
-
-if uploaded_file is not None:
-    # Lendo o arquivo
-    if uploaded_file.name.endswith('.csv'):
-        df = pd.read_csv(uploaded_file)
-    else:
-        df = pd.read_excel(uploaded_file)
-
-    # Verificando as colunas necessárias
-    colunas_necessarias = ["Custo Produto", "Taxa Imposto", "Taxa Shopee", "Valor Fixo Shopee", "Lucro Esperado", "Lucro Minimo"]
-    if all(coluna in df.columns for coluna in colunas_necessarias):
-        # Processando a planilha
-        df_resultados = utils.processar_planilha(df)
-
-        # Exibindo os resultados
-        st.write("### Resultados")
-        st.dataframe(df_resultados)
-
-        # Botão para exportar os resultados para Excel
-        output = BytesIO()
-        with pd.ExcelWriter(output, engine="xlsxwriter") as writer:
-            df_resultados.to_excel(writer, index=False, sheet_name="Resultados")
-        output.seek(0)
-
-        st.download_button(
-            label="Exportar para Excel",
-            data=output,
-            file_name="resultados.xlsx",
-            mime="application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
+# Função para processar a planilha
+def processar_planilha(df, taxa_imposto, taxa_shopee, valor_fixo_shopee, considera_operacional, faturamento_mensal, custos_fixos, lucro_esperado, lucro_minimo):
+    resultados = []
+    for index, row in df.iterrows():
+        # Criando uma instância da classe para cada produto
+        calculadora = CalculoPrecoShopee(
+            custo_produto = row["Custo Produto"],
+            custo_embalagem = row.get("Custo Embalagem", 0),  # Valor padrão se a coluna não existir
+            taxa_imposto = taxa_imposto,
+            taxa_shopee = taxa_shopee,
+            valor_fixo_shopee = valor_fixo_shopee,
+            outros_custos = row.get("Outros Custos", 0),
+            considera_operacional = considera_operacional,
+            faturamento_mensal = faturamento_mensal,
+            custos_fixos = custos_fixos,
+            lucro_esperado = lucro_esperado,
+            lucro_minimo = lucro_minimo
         )
-    else:
-        st.error(f"A planilha deve conter as colunas: {', '.join(colunas_necessarias)}")
-else:
-    st.info("Faça o upload de uma planilha para começar.")
+
+        # Calculando os valores
+        calculadora.calcula_preco_venda()
+
+        valor_comissao_shopee, perc_comissao_shopee, valor_imposto, receita_operacional = calculadora.calcula_comissoes_impostos_receita('final')
+        lucro = calculadora.calcula_lucro('final')
+
+        sugestao_aumento = calculadora.sugere_aumento()
+
+        # Armazenando os resultados
+        resultados.append({
+            "Nome Produto": row.get("Nome Produto", f"Produto {index + 1}"),  # Nome do produto (se existir)
+            "Preço de Venda": calculadora.preco_ideal,
+            "Lucro (R$)": lucro,
+            "Desconto Máximo": calculadora.max_desconto * 100,  # Em porcentagem
+            "Comissão Shopee (R$)": valor_comissao_shopee,
+            "% Comissão Shopee": perc_comissao_shopee * 100,  # Em porcentagem
+            "Valor Imposto (R$)": valor_imposto,
+            "Receita Operacional (R$)": receita_operacional,
+            "Sugestão de Aumento": sugestao_aumento
+        })
+
+    return pd.DataFrame(resultados)
