+++ src/utils.py
@@ -1,11 +1,9 @@
 import streamlit as st
 import sys
 import os
-import pandas as pd
 
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 from src import config
-from classes.CalculoPrecoShopee import CalculoPrecoShopee
 
 # Função para pegar parâmetros do Streamlit
 def define_campos_inputs_principal():
@@ -97,46 +95,3 @@ def define_campos_inputs_massa():
 def load_css(file_path):
     with open(file_path, "r") as f:
         st.markdown(f"<style>{f.read()}</style>", unsafe_allow_html=True)
-
-
-# Função para processar a planilha
-def processar_planilha(df, taxa_imposto, taxa_shopee, valor_fixo_shopee, considera_operacional, faturamento_mensal, custos_fixos, lucro_esperado, lucro_minimo):
-    resultados = []
-    for index, row in df.iterrows():
-        # Criando uma instância da classe para cada produto
-        calculadora = CalculoPrecoShopee(
-            custo_produto = row["Custo Produto"],
-            custo_embalagem = row.get("Custo Embalagem", 0),  # Valor padrão se a coluna não existir
-            taxa_imposto = taxa_imposto,
-            taxa_shopee = taxa_shopee,
-            valor_fixo_shopee = valor_fixo_shopee,
-            outros_custos = row.get("Outros Custos", 0),
-            considera_operacional = considera_operacional,
-            faturamento_mensal = faturamento_mensal,
-            custos_fixos = custos_fixos,
-            lucro_esperado = lucro_esperado,
-            lucro_minimo = lucro_minimo
-        )
-
-        # Calculando os valores
-        calculadora.calcula_preco_venda()
-
-        valor_comissao_shopee, perc_comissao_shopee, valor_imposto, receita_operacional = calculadora.calcula_comissoes_impostos_receita('final')
-        lucro = calculadora.calcula_lucro('final')
-
-        sugestao_aumento = calculadora.sugere_aumento()
-
-        # Armazenando os resultados
-        resultados.append({
-            "Nome Produto": row.get("Nome Produto", f"Produto {index + 1}"),  # Nome do produto (se existir)
-            "Preço de Venda": calculadora.preco_ideal,
-            "Lucro (R$)": lucro,
-            "Desconto Máximo": calculadora.max_desconto * 100,  # Em porcentagem
-            "Comissão Shopee (R$)": valor_comissao_shopee,
-            "% Comissão Shopee": perc_comissao_shopee * 100,  # Em porcentagem
-            "Valor Imposto (R$)": valor_imposto,
-            "Receita Operacional (R$)": receita_operacional,
-            "Sugestão de Aumento": sugestao_aumento
-        })
-
-    return pd.DataFrame(resultados)
